fix: connection을 연도별로 나눠 관리하기
