+++ home/src/features/connections/fixtures/2022/index.ts
@@ -1,0 +1,533 @@
+import { ConnectionsType } from "..";
+
+export const CONNECTIONS2022: ConnectionsType[] = [
+  {
+    week: 1,
+    quiz: [
+      {
+        words: ["보석", "수박", "캔디", "돼지"],
+        description: "아이스크림 __바",
+      },
+      {
+        words: ["구리", "용인", "화성", "양주"],
+        description: "경기도의 행정구역",
+      },
+      {
+        words: ["별똥별", "지진", "일식", "번개"],
+        description: "자연현상",
+      },
+      {
+        words: ["더위", "신사", "관심", "소유"],
+        description: "무__",
+      },
+    ],
+  },
+
+  {
+    week: 2,
+    quiz: [
+      {
+        words: ["수영", "체조", "유도", "조정"],
+        description: "올림픽 종목",
+      },
+      {
+        words: ["돌", "콩", "숙주", "도라지"],
+        description: "__나물",
+      },
+      {
+        words: ["괴물", "기생충", "해운대", "암살"],
+        description: "대한민국의 천만 관객 영화",
+      },
+      {
+        words: ["복수", "다양", "다중", "곱절"],
+        description: "여러개의",
+      },
+    ],
+  },
+  {
+    week: 3,
+    quiz: [
+      {
+        words: ["꽃", "가지", "뿌리", "열매"],
+        description: "식물의 기관",
+      },
+      {
+        words: ["주사", "겸자", "핀셋", "가위"],
+        description: "수술도구",
+      },
+      {
+        words: ["문신", "볼펜", "만년필", "프린터"],
+        description: "잉크를 사용하는 것",
+      },
+      {
+        words: ["게", "커튼", "와이퍼", "자동문"],
+        description: "옆으로만 움직이는 것",
+      },
+    ],
+  },
+  {
+    week: 4,
+    quiz: [
+      {
+        words: ["자두", "살구", "복숭아", "아보카도"],
+        description: "하나의 큰 씨가 있는 과일",
+      },
+      {
+        words: ["고추장", "케첩", "살사", "베샤멜"],
+        description: "소스",
+      },
+      {
+        words: ["DNA", "다이너마이트", "버터", "아이돌"],
+        description: "BTS의 노래 제목",
+      },
+      {
+        words: ["슈퍼", "포터", "배트", "팩"],
+        description: "___맨",
+      },
+    ],
+  },
+  {
+    week: 5,
+    quiz: [
+      {
+        words: ["위", "간", "이자", "기도"],
+        description: "신체 기관",
+      },
+      {
+        words: ["장구", "북", "공", "오르골"],
+        description: "타악기",
+      },
+      {
+        words: ["시트러스", "머스크", "우디", "허브"],
+        description: "향수의 향",
+      },
+      {
+        words: ["갈라파고스", "솔로몬", "쿡", "카나리아"],
+        description: "제도",
+      },
+    ],
+  },
+  {
+    week: 6,
+    quiz: [
+      {
+        words: ["홀덤", "방탈출", "퀴즈", "소설"],
+        description: "추러스 활동",
+      },
+      {
+        words: ["얼음", "김", "눈", "강"],
+        description: "물의 형태",
+      },
+      {
+        words: ["구설수", "왕위", "산", "궤도"],
+        description: "오를 수 있는 것",
+      },
+      {
+        words: ["컵", "비행기", "빨대", "호랑이"],
+        description: "종이 ___",
+      },
+    ],
+  },
+  {
+    week: 7,
+    quiz: [
+      {
+        words: ["중국", "스페인", "아이티", "수단"],
+        description: "국가",
+      },
+      {
+        words: ["디스코", "스트릿", "포크", "포고"],
+        description: "춤",
+      },
+      {
+        words: ["로퍼", "하이힐", "옥스퍼드", "메리제인"],
+        description: "구두",
+      },
+      {
+        words: ["골", "윙", "크로스", "티키타카"],
+        description: "축구 용어",
+      },
+    ],
+  },
+  {
+    week: 8,
+    quiz: [
+      {
+        words: ["휘발유", "아스팔트", "비닐", "플라스틱"],
+        description: "석유를 원료로 하는 것",
+      },
+      {
+        words: ["비", "피", "오", "아이"],
+        description: "알파벳",
+      },
+      {
+        words: ["붓꽃", "해바라기", "자화상", "모나리자"],
+        description: "명화",
+      },
+      {
+        words: ["심판", "달", "연인", "악마"],
+        description: "타로카드",
+      },
+    ],
+  },
+  {
+    week: 9,
+    quiz: [
+      {
+        words: ["척추", "해골", "갈비", "골반"],
+        description: "인체의 뼈",
+      },
+      {
+        words: ["상아", "진주", "호박", "산호"],
+        description: "유기물로 만들어진 보석",
+      },
+      {
+        words: ["톳", "보루", "판", "쌍"],
+        description: "한국의 단위",
+      },
+      {
+        words: ["마지막", "밥", "손", "홀"],
+        description: "맨___",
+      },
+    ],
+  },
+  {
+    week: 10,
+    quiz: [
+      {
+        words: ["개", "토끼", "원숭이", "용"],
+        description: "십이지신",
+      },
+      {
+        words: ["벽록", "군청", "밤", "갈매"],
+        description: "색",
+      },
+      {
+        words: ["고름", "깃", "소매", "폭"],
+        description: "한복의 부분",
+      },
+      {
+        words: ["생존", "그림", "유전", "화학"],
+        description: "___자",
+      },
+    ],
+  },
+  {
+    week: 11,
+    quiz: [
+      {
+        words: ["무좀", "사마귀", "건선", "티눈"],
+        description: "피부 질환",
+      },
+      {
+        words: ["버섯", "효모", "곰팡이", "동충하초"],
+        description: "균류",
+      },
+      {
+        words: ["비행기", "새", "파리", "선풍기"],
+        description: "날개가 있는 것",
+      },
+      {
+        words: ["전화기", "TV", "텔레파시", "망원경"],
+        description: "영어가 tele___ 인 단어",
+      },
+    ],
+  },
+  {
+    week: 12,
+    quiz: [
+      {
+        words: ["주석", "납", "비소", "수은"],
+        description: "금속 원소",
+      },
+      {
+        words: ["슬림", "스키니", "테이퍼드", "와이드"],
+        description: "바지 핏",
+      },
+      {
+        words: ["벤치", "짐볼", "아령", "랙"],
+        description: "운동기구",
+      },
+      {
+        words: ["아이언", "우드", "웨지", "퍼터"],
+        description: "골프클럽",
+      },
+    ],
+  },
+  {
+    week: 13,
+    quiz: [
+      {
+        words: ["포", "기갑", "총", "리볼버"],
+        description: "현대 무기",
+      },
+      {
+        words: ["루피", "패티", "로디", "해리"],
+        description: "뽀로로의 캐릭터",
+      },
+      {
+        words: ["프랑", "루블", "원", "페소"],
+        description: "통화 단위",
+      },
+      {
+        words: ["푹스", "파인만", "노이만", "보어"],
+        description: "맨해튼 프로젝트의 과학자",
+      },
+    ],
+  },
+  {
+    week: 14,
+    quiz: [
+      {
+        words: ["뱀", "코끼리", "표범", "사자"],
+        description: "바다___ 인 동물",
+      },
+      {
+        words: ["기업", "우리", "하나", "국민"],
+        description: "___은행",
+      },
+      {
+        words: ["동해", "남산", "가을", "기상"],
+        description: "애국가 첫 단어",
+      },
+      {
+        words: ["아파트", "단지", "번지", "호수"],
+        description: "주소",
+      },
+    ],
+  },
+  {
+    week: 15,
+    quiz: [
+      {
+        words: ["강원", "라라", "서울", "디즈니"],
+        description: "___랜드",
+      },
+      {
+        words: ["신", "열", "참깨", "틈새"],
+        description: "___라면",
+      },
+      {
+        words: ["탑", "바다", "유리", "길"],
+        description: "연예인",
+      },
+      {
+        words: ["솔", "파라솔", "도시", "파도"],
+        description: "계이름으로 이루어진 단어",
+      },
+    ],
+  },
+  {
+    week: 16,
+    quiz: [
+      {
+        words: ["캐모마일", "루이보스", "히비스커스", "페퍼민트"],
+        description: "차",
+      },
+
+      {
+        words: ["스매시", "커트", "서브", "드라이브"],
+        description: "탁구 용어",
+      },
+      {
+        words: ["시스루", "보브", "리프", "바가지"],
+        description: "헤어스타일",
+      },
+      {
+        words: ["올리브", "코발트", "크림슨", "아이보리"],
+        description: "색",
+      },
+    ],
+  },
+  {
+    week: 17,
+    quiz: [
+      {
+        words: ["스톡홀름", "마드리드", "오슬로", "코펜하겐"],
+        description: "유럽 국가의 수도",
+      },
+      {
+        words: ["안단테", "알레그로", "라르고", "비바체"],
+        description: "음악 빠르기 용어",
+      },
+      {
+        words: ["이오", "유로파", "칼리스토", "가니메데"],
+        description: "목성의 갈릴레이 위성",
+      },
+      {
+        words: ["와칸다", "소코비아", "뉴욕", "타이탄"],
+        description: "어벤져스 전투 지역",
+      },
+    ],
+  },
+  {
+    week: 18,
+    quiz: [
+      {
+        words: ["철학", "종교", "언어", "미학"],
+        description: "인문학 분야",
+      },
+      {
+        words: ["사냥", "낚시", "채집", "농사"],
+        description: "먹을 것을 얻는 방법",
+      },
+      {
+        words: ["지레", "미리", "사전", "먼저"],
+        description: "시간상으로 이전의",
+      },
+      {
+        words: ["문어", "네모", "후라이", "나비"],
+        description: "___의 꿈 (노래)",
+      },
+    ],
+  },
+  {
+    week: 19,
+    quiz: [
+      {
+        words: ["비트", "근대", "사탕무", "시금치"],
+        description: "비름과 채소",
+      },
+      {
+        words: ["킹", "더블", "싱글", "퀸"],
+        description: "침대 사이즈",
+      },
+      {
+        words: ["탭", "버즈", "워치", "노트"],
+        description: "삼성 갤럭시 제품 시리즈",
+      },
+      {
+        words: ["커피", "박스", "스터디", "바이트"],
+        description: "메가___",
+      },
+    ],
+  },
+  {
+    week: 20,
+    quiz: [
+      {
+        words: ["시장", "허기", "공복", "출출"],
+        description: "배고픔",
+      },
+      {
+        words: ["단풍", "싸리", "모란", "국화"],
+        description: "화투 패의 그림",
+      },
+      {
+        words: ["묘수", "불계", "무리수", "포석"],
+        description: "바둑 용어",
+      },
+      {
+        words: ["향유", "돌", "북극", "대왕"],
+        description: "___고래",
+      },
+    ],
+  },
+  {
+    week: 21,
+    quiz: [
+      {
+        words: ["북극", "대서", "태평", "인도"],
+        description: "바다",
+      },
+      {
+        words: ["백로", "하지", "청명", "대설"],
+        description: "24절기",
+      },
+      {
+        words: ["점검", "검토", "조사", "검정"],
+        description: "분석하여 따짐",
+      },
+      {
+        words: ["소개", "관리", "계발", "자랑"],
+        description: "자기___",
+      },
+    ],
+  },
+  {
+    week: 22,
+    quiz: [
+      {
+        words: ["샤베트", "젤라토", "하드", "슬러쉬"],
+        description: "빙과류 디저트",
+      },
+      {
+        words: ["페어", "탑", "트리플", "스트레이트"],
+        description: "포커 족보",
+      },
+
+      {
+        words: ["카지노", "루시", "파이란", "올드보이"],
+        description: "최민식 출연 영화",
+      },
+
+      {
+        words: ["리프팅", "필러", "보톡스", "인모드"],
+        description: "피부과 시술",
+      },
+    ],
+  },
+  {
+    week: 23,
+    quiz: [
+      {
+        words: ["부아", "화", "골", "노여움"],
+        description: "분노",
+      },
+      {
+        words: ["쇄신", "개벽", "혁명", "변경"],
+        description: "개혁",
+      },
+      {
+        words: ["맹장", "충수", "직장", "결장"],
+        description: "큰 창자의 부위",
+      },
+      {
+        words: ["관수", "개관", "관개", "급수"],
+        description: "논밭에 물을 대는 것",
+      },
+    ],
+  },
+  {
+    week: 24,
+    quiz: [
+      {
+        words: ["마늘", "와사비", "마라", "겨자"],
+        description: "매운맛의 소스",
+      },
+      {
+        words: ["고수", "공유", "주원", "현빈"],
+        description: "한국 남자배우",
+      },
+      {
+        words: ["대추", "사과", "포", "밤"],
+        description: "차례상 음식",
+      },
+      {
+        words: ["콩", "소시지", "햄", "사리"],
+        description: "부대찌개 재료",
+      },
+    ],
+  },
+  {
+    week: 25,
+    quiz: [
+      {
+        words: ["철인", "군자", "신사", "히어로"],
+        description: "훌륭한 사람",
+      },
+      {
+        words: ["불", "바람", "꽃", "게으름"],
+        description: "피울 수 있는 것",
+      },
+      {
+        words: ["보살", "포일", "배터리", "배트"],
+        description: "야구 용어",
+      },
+      {
+        words: ["볼", "강아지", "인재", "물"],
+        description: "___풀",
+      },
+    ],
+  },
+];
