+++ home/src/features/connections/fixtures/2024/index.ts
@@ -1,0 +1,25 @@
+import { ConnectionsType } from "..";
+
+export const CONNECTIONS2023: ConnectionsType[] = [
+  {
+    week: 1,
+    quiz: [
+      {
+        words: ["블랙", "노랑", "보라", "백"],
+        description: "색깔",
+      },
+      {
+        words: ["차", "철도", "신호등", "택시"],
+        description: "교통과 관련된 단어",
+      },
+      {
+        words: ["보리", "기장", "파", "백도"],
+        description: "식물",
+      },
+      {
+        words: ["도", "만", "보다", "가"],
+        description: "조사",
+      },
+    ],
+  },
+];
